requirements.txt
# Web scraping
beautifulsoup4==4.12.3
lxml==5.1.0
markdownify==0.12.1
playwright==1.48.0
requests==2.31.0

# Database
supabase==2.9.0

# Utilities
python-dotenv==1.0.1
python-dateutil==2.8.2
